Move node-cli within node
