+++ Cargo.toml
@@ -2,6 +2,10 @@
 name = "substrate"
 path = "node/src/main.rs"
 
+[lib]
+name = "substrate_node_cli"
+path = "node/src/lib.rs"
+
 [package]
 name = "substrate"
 version = "2.0.0"
@@ -10,11 +14,50 @@ build = "build.rs"
 edition = "2018"
 
 [dependencies]
-cli = { package = "node-cli", path = "node/cli" }
+log = "0.4"
+tokio = "0.1.7"
 futures = "0.1"
 ctrlc = { version = "3.0", features = ["termination"] }
+exit-future = "0.1"
+cli = { package = "substrate-cli", path = "core/cli" }
+parity-codec = { version = "4.1.1" }
+sr-io = { path = "core/sr-io" }
+client = { package = "substrate-client", path = "core/client" }
+primitives = { package = "substrate-primitives", path = "core/primitives" }
+inherents = { package = "substrate-inherents", path = "core/inherents" }
+node-runtime = { path = "node/runtime" }
+node-primitives = { path = "node/primitives" }
+hex-literal = "0.2"
+substrate-basic-authorship = { path = "core/basic-authorship" }
+substrate-service = { path = "core/service" }
+transaction_pool = { package = "substrate-transaction-pool", path = "core/transaction-pool" }
+network = { package = "substrate-network", path = "core/network" }
+aura = { package = "substrate-consensus-aura", path = "core/consensus/aura" }
+aura_primitives = { package = "substrate-consensus-aura-primitives", path = "core/consensus/aura/primitives" }
+grandpa = { package = "substrate-finality-grandpa", path = "core/finality-grandpa" }
+grandpa_primitives = { package = "substrate-finality-grandpa-primitives", path = "core/finality-grandpa/primitives" }
+sr-primitives = { path = "core/sr-primitives" }
+node-executor = { path = "node/executor" }
+substrate-keystore = { path = "core/keystore" }
+substrate-telemetry = { package = "substrate-telemetry", path = "core/telemetry" }
+structopt = "0.2"
+transaction-factory = { path = "test-utils/transaction-factory" }
+keyring = { package = "substrate-keyring", path = "core/keyring" }
+indices = { package = "srml-indices", path = "srml/indices" }
+timestamp = { package = "srml-timestamp", path = "srml/timestamp", default-features = false }
+rand = "0.6"
+finality_tracker = { package = "srml-finality-tracker", path = "srml/finality-tracker", default-features = false }
+contracts = { package = "srml-contracts", path = "srml/contracts" }
+system = { package = "srml-system", path = "srml/system" }
+balances = { package = "srml-balances", path = "srml/balances" }
+
+[dev-dependencies]
+consensus-common = { package = "substrate-consensus-common", path = "core/consensus/common" }
+service-test = { package = "substrate-service-test", path = "core/service/test" }
 
 [build-dependencies]
+cli = { package = "substrate-cli", path = "core/cli" }
+structopt = "0.2"
 vergen = "3"
 
 [workspace]
@@ -86,7 +129,6 @@ members = [
 	"srml/system",
 	"srml/timestamp",
 	"srml/treasury",
-	"node/cli",
 	"node/executor",
 	"node/primitives",
 	"node/runtime",
