--- node/cli/Cargo.toml
@@ -1,52 +1,0 @@
-[package]
-name = "node-cli"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-description = "Substrate node implementation in Rust."
-build = "build.rs"
-edition = "2018"
-
-[dependencies]
-log = "0.4"
-tokio = "0.1.7"
-futures = "0.1"
-exit-future = "0.1"
-cli = { package = "substrate-cli", path = "../../core/cli" }
-parity-codec = { version = "4.1.1" }
-sr-io = { path = "../../core/sr-io" }
-client = { package = "substrate-client", path = "../../core/client" }
-primitives = { package = "substrate-primitives", path = "../../core/primitives" }
-inherents = { package = "substrate-inherents", path = "../../core/inherents" }
-node-runtime = { path = "../runtime" }
-node-primitives = { path = "../primitives" }
-hex-literal = "0.2"
-substrate-basic-authorship = { path = "../../core/basic-authorship" }
-substrate-service = { path = "../../core/service" }
-transaction_pool = { package = "substrate-transaction-pool", path = "../../core/transaction-pool" }
-network = { package = "substrate-network", path = "../../core/network" }
-aura = { package = "substrate-consensus-aura", path = "../../core/consensus/aura" }
-aura_primitives = { package = "substrate-consensus-aura-primitives", path = "../../core/consensus/aura/primitives" }
-grandpa = { package = "substrate-finality-grandpa", path = "../../core/finality-grandpa" }
-grandpa_primitives = { package = "substrate-finality-grandpa-primitives", path = "../../core/finality-grandpa/primitives" }
-sr-primitives = { path = "../../core/sr-primitives" }
-node-executor = { path = "../executor" }
-substrate-keystore = { path = "../../core/keystore" }
-substrate-telemetry = { package = "substrate-telemetry", path = "../../core/telemetry" }
-structopt = "0.2"
-transaction-factory = { path = "../../test-utils/transaction-factory" }
-keyring = { package = "substrate-keyring", path = "../../core/keyring" }
-indices = { package = "srml-indices", path = "../../srml/indices" }
-timestamp = { package = "srml-timestamp", path = "../../srml/timestamp", default-features = false }
-rand = "0.6"
-finality_tracker = { package = "srml-finality-tracker", path = "../../srml/finality-tracker", default-features = false }
-contracts = { package = "srml-contracts", path = "../../srml/contracts" }
-system = { package = "srml-system", path = "../../srml/system" }
-balances = { package = "srml-balances", path = "../../srml/balances" }
-
-[dev-dependencies]
-consensus-common = { package = "substrate-consensus-common", path = "../../core/consensus/common" }
-service-test = { package = "substrate-service-test", path = "../../core/service/test" }
-
-[build-dependencies]
-cli = { package = "substrate-cli", path = "../../core/cli" }
-structopt = "0.2"
