+++ test-utils/chain-spec-builder/Cargo.toml
@@ -6,6 +6,6 @@ edition = "2018"
 
 [dependencies]
 clap = { version = "~2.32", features = ["yaml"] }
-node-cli = { path = "../../node/cli" }
+substrate = { path = "../.." }
 substrate-primitives = { path = "../../core/primitives" }
 substrate-service = { path = "../../core/service" }
